requirements.txt
psycopg2~=2.9.6
Django~=4.2
djangorestframework~=3.14.0

